requirements.txt
fastapi 
strawberry-graphql[fastapi] 
sqlalchemy 
asyncpg 
uvicorn
python-dotenv